不同表修改，添加createService.sync

+++ src/main/java/org/corefine/mysqlsync/service/CreateService.java
@@ -9,30 +9,29 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class CreateService {
-	private final Logger logger = LoggerFactory.getLogger(getClass());
-	@Autowired
-	private DBService dbService;
-	@Autowired
-	private Md5Service md5Service;
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+    @Autowired
+    private DBService dbService;
+    @Autowired
+    private Md5Service md5Service;
 
-	public void sync(SyncConnection syncConnection, DatabaseSyncConfig db) {
-//		String sql = "show create table " + db.getSrcTableName();
-//		String srcData = dbService.queryOne(syncConnection.src, sql).get("Create Table").toString();
-//		String srcMd5 = md5Service.md5(srcData);
-		String descData = "";
-		boolean isDelete;
-		try {
-			String sql = "show create table " + db.getDescTableName();
-			descData = dbService.queryOne(syncConnection.desc, sql).get("Create Table").toString();
-			isDelete = true;
-		} catch (RuntimeException e) {
-			isDelete = false;
-			//table is not exist
-		}
-		if (isDelete)
-			dbService.execute(syncConnection.desc, "drop table " + db.getDescTableName());
-		System.out.println("descData:" + descData);
-		dbService.execute(syncConnection.desc, descData);
-		logger.info(db.getDescTableName() + "表结构错误，已经重建");
-	}
+    public void sync(SyncConnection syncConnection, DatabaseSyncConfig db) {
+        boolean isDelete;
+        String srcData = null;
+        String descData = null;
+        try {
+            srcData = dbService.queryOne(syncConnection.src, "show create table " + db.getSrcTableName()).get("Create Table").toString();
+            srcData = srcData.replaceFirst(db.getSrcTableName(), db.getDescTableName());
+            descData = dbService.queryOne(syncConnection.desc, "show create table " + db.getDescTableName()).get("Create Table").toString();
+            isDelete = true;
+        } catch (RuntimeException e) {
+            isDelete = false;
+        }
+        if (!md5Service.md5(srcData).equals(md5Service.md5(descData))) {
+            if (isDelete)
+                dbService.execute(syncConnection.desc, "drop table " + db.getDescTableName());
+            dbService.execute(syncConnection.desc, srcData);
+            logger.info(db.getDescTableName() + "表结构错误，已经重建");
+        }
+    }
 }
